% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{all_rooms}
\alias{all_rooms}
\title{Get the room IDs of all joined rooms.}
\usage{
all_rooms(initial_sync)
}
\arguments{
\item{initial_sync}{The result of a call to \code{\link[=sync]{sync()}}. This will be used to
find the joined rooms.}
}
\value{
A character vector of room IDs.
}
\description{
Get the room IDs of all joined rooms.
}
