% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_members}
\alias{get_members}
\title{Retrieve a list of memberships for one Matrix room.}
\usage{
get_members(room_id)
}
\arguments{
\item{room_id}{The room to receive members for.}
}
\value{
A list object from the Matrix API.
}
\description{
This will fetch all memberships for a given room at the current point in
time. Note that this will include currently joined users as well as all users
that have been in the room historically. They can be differentiated by their
\code{membership} field.
}
