% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_messages}
\alias{get_messages}
\title{Retrieve messages using the Matrix-Client-Server API.}
\usage{
get_messages(room_id, from, dir = "b", to = NULL)
}
\arguments{
\item{room_id}{The room whose messages will be received.}

\item{from}{The start token within the history.}

\item{dir}{Direction to receive messages in. "b" for backwards in time, "f"
for forwards in time.}

\item{to}{The end token to stop receiving messages at.}
}
\value{
A list object from the Matrix API.
}
\description{
Retrieve messages using the Matrix-Client-Server API.
}
