% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room_history.R
\name{get_rooms}
\alias{get_rooms}
\title{Get the room events for a given room ID.}
\usage{
get_rooms(room_ids, since, sync = NULL)
}
\arguments{
\item{room_ids}{Vector of room IDs to iterate over.}

\item{since}{Stop paginating when reaching this time.}

\item{sync}{(Optional) Result of a prior call to \code{\link[=sync]{sync()}} to save
duplication.}
}
\value{
A tibble containing event information.
}
\description{
Get the room events for a given room ID.
}
