% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{initial_sync}
\alias{initial_sync}
\title{Perform an initial sync using the Matrix Client-Server API.}
\usage{
initial_sync(since = NULL)
}
\arguments{
\item{since}{Optional parameter to continue from a previous sync.}
}
\value{
A list object from the Matrix API.
}
\description{
Perform an initial sync using the Matrix Client-Server API.
}
