% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_room_events.R
\name{parse_room_events}
\alias{parse_room_events}
\title{Parse the JSON list for a room into a rectangular format}
\usage{
parse_room_events(event_list)
}
\arguments{
\item{event_list}{the list of events from the Matrix API}
}
\value{
A table of event data
}
\description{
Parse the JSON list for a room into a rectangular format
}
