% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_active_times}
\alias{plot_active_times}
\title{Create a plot showing the most active hours across the provided rooms.}
\usage{
plot_active_times(rooms)
}
\arguments{
\item{rooms}{Rooms object to use. See \code{\link[=rooms]{rooms()}}.}
}
\value{
A \code{ggplot2} object for plotting.
}
\description{
Create a plot showing the most active hours across the provided rooms.
}
