% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_daily_activity}
\alias{plot_daily_activity}
\title{Create a plot showing the daily activity across the provided rooms.}
\usage{
plot_daily_activity(rooms)
}
\arguments{
\item{rooms}{Rooms object to use. See \code{\link[=rooms]{rooms()}}.}
}
\value{
A \code{ggplot2} object for plotting.
}
\description{
Create a plot showing the daily activity across the provided rooms.
}
