% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_user_network}
\alias{plot_user_network}
\title{Create a network plot displaying the adjacency of users.}
\usage{
plot_user_network(
  rooms,
  joined_threshold = 2,
  user_blacklist = NULL,
  shared_threshold = 2,
  show_labels = TRUE,
  label_threshold = 5,
  label_whitelist = NULL
)
}
\arguments{
\item{rooms}{Rooms object to use. See \code{\link[=rooms]{rooms()}}.}

\item{joined_threshold}{Minimum number of joined rooms for including a user.}

\item{user_blacklist}{User IDs that should be excluded.}

\item{shared_threshold}{Minimum number of shared rooms for including a
connection between two users. The default value is 2 to accomodate the
fact, that all users share a room with the user doing the request.}

\item{show_labels}{Whether to show labels. This parameter does not apply to
the users on the \code{label_whitelist}.}

\item{label_threshold}{Minimum number of joined rooms for labeling a user.}

\item{label_whitelist}{User IDs that should be labeled.}
}
\value{
An \code{igraph} object for plotting. See \link[igraph:plot.common]{igraph::igraph.plotting} for
information on tweaking the visuals.
}
\description{
The network graph is constructed by putting users into close proximity based
on their shared joined rooms.
}
\examples{

# Retrieve some rooms (in this case all joined rooms):

initial_sync <- sync()

rooms <- get_rooms(
  all_rooms(initial_sync),
  since = lubridate::today(),
  initial_sync
)

# Create the network graph:

graph <- plot_user_network(rooms)

# Plot the graph tweaking some of the visuals:

plot(
  graph,
  edge.color = "#BBBBBB1A",
  vertex.frame.color = NA,
  vertex.color = "#303030AA"
)

}
