% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{print.ChatStat_rooms}
\alias{print.ChatStat_rooms}
\title{Print information on a rooms object.}
\usage{
\method{print}{ChatStat_rooms}(x, ...)
}
\arguments{
\item{x}{The rooms object to print.}

\item{...}{Ignored, for compatability with print.default}
}
\description{
Print information on a rooms object.
}
