% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_report}
\alias{render_report}
\title{Render an HTML report based on a rooms object.}
\usage{
render_report(rooms, output_file = NULL)
}
\arguments{
\item{rooms}{Rooms object to base the report on. See \code{\link[=rooms]{rooms()}}.}

\item{output_file}{The place to write the report to. Defaults to
\code{chatstat_report.html}.}
}
\description{
Please note, that the current working directory has to be writable in order
to save the output as well as the intermediate files.
}
