% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room_history.R
\name{room_history}
\alias{room_history}
\title{Get the list of room events for a given room ID.}
\usage{
room_history(room_id, since = NULL)
}
\arguments{
\item{room_id}{The room to get data for.}

\item{since}{Stop paginating when reaching this time.}
}
\value{
A list of events.
}
\description{
Get the list of room events for a given room ID.
}
