% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room.R
\name{update_room}
\alias{update_room}
\title{Update the room state using the Matrix API.}
\usage{
update_room(room, sync = NULL)
}
\arguments{
\item{room}{The room to update.}

\item{sync}{The result of a recent call to \code{\link[=sync]{sync()}}. This will be called
automatically, if the parameter is missing. However, it is recommended to
share the sync between multiple rooms to avoid duplicated work and to have
a common base for comparing the timelines.}
}
\description{
This function extends the existing state of the room with new events from
the Matrix API. Events are added from a recent call to \code{\link[=sync]{sync()}} and the gap
to the existing timeline is filled with additional calls to the messages
API.
}
