% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rooms.R
\name{update_rooms}
\alias{update_rooms}
\title{Update room state using the Matrix API.}
\usage{
update_rooms(rooms, sync = NULL)
}
\arguments{
\item{rooms}{The rooms to update. See \code{\link[=rooms]{rooms()}}.}

\item{sync}{The result of a recent call to \code{\link[=sync]{sync()}}. This will be called
automatically, if the parameter is missing. Note, that this should be a
sync that was done using the \code{next_token} item from the rooms object as
the \code{since} parameter.}
}
\description{
This function extends the existing state of the rooms with new events from
the Matrix API. Events are added from a recent call to \code{\link[=sync]{sync()}} and the gap
to the existing timeline is filled with additional calls to the messages
API.
}
