% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{get_messages}
\alias{get_messages}
\title{Retrieve messages using the Matrix-Client-Server API.}
\usage{
get_messages(room_id, from)
}
\arguments{
\item{room_id}{The room whose messages will be received.}

\item{from}{The start token within the history.}
}
\value{
A list object from the Matrix API.
}
\description{
Retrieve messages using the Matrix-Client-Server API.
}
