% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_report}
\alias{render_report}
\title{Render a basic HTML report based on a set of room data}
\usage{
render_report(room_data, output_file = "./chatstat_report.html")
}
\arguments{
\item{room_data}{A tibble of messages. Expects to be unpacked from
the nested format in get_rooms}

\item{output_file}{The place to write the report. Defaults
to \code{./chatstat_report.html}}
}
\value{
A list object from the Matrix API.
}
\description{
Render a basic HTML report based on a set of room data
}
