% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_report}
\alias{render_report}
\title{Render an HTML report based on room events.}
\usage{
render_report(events, output_file = NULL)
}
\arguments{
\item{events}{Events to include in the report. This should be from a \code{rooms}
object's \code{events} field.}

\item{output_file}{The place to write the report to. Defaults to
\code{chatstat_report.html}.}
}
\description{
Please note, that the current working directory has to be writable in order
to save the output as well as the intermediate files.
}
