% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room_history.R
\name{room_history}
\alias{room_history}
\title{Get the list of room events for a given roomID}
\usage{
room_history(roomid, since = NULL)
}
\arguments{
\item{roomid}{the room to get data for}

\item{since}{stop paginating when reaching this time}
}
\value{
A list of events
}
\description{
Get the list of room events for a given roomID
}
