% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync}
\alias{sync}
\title{Call the initial sync from the Matrix Client-Server API}
\usage{
sync(since = NULL)
}
\arguments{
\item{since}{Optional parameter to continue from a previous sync}
}
\value{
A list object from the Matrix API
}
\description{
Call the initial sync from the Matrix Client-Server API
}
